src/requirements.txt
amqp==5.2.0
aniso8601==9.0.1
asgiref==3.7.2
billiard==4.2.0
celery==5.3.6
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cron-descriptor==1.4.3
Django==5.0.2
django-celery-beat==2.6.0
django-cors-headers==4.3.1
django-timezone-field==6.1.0
flower==2.0.1
graphene==3.3
graphene-django==3.2.0
graphql-core==3.2.3
graphql-relay==3.2.0
humanize==4.9.0
idna==3.6
kombu==5.3.5
nba_api==1.4.1
numpy==1.26.4
pandas==2.2.1
pillow==10.4.0
prometheus_client==0.20.0
promise==2.3
prompt-toolkit==3.0.43
python-crontab==3.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
redis==5.0.3
redis-cli==1.0.1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
text-unidecode==1.3
tornado==6.4
tzdata==2024.1
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
# windows-curses==2.3.2